refactor individual scripts into a module
