--- python-neo4j-driver/neo4j_config.py
@@ -1,6 +1,0 @@
-import yaml
-
-def neo4jConfig(configFile):
-    with open(configFile, "r") as ymlfile:
-        cfg = yaml.safe_load(ymlfile)
-    return cfg
